Added new output column and fixed choice made
Choice made was actually unnecessary, so it was removed
from utils. Some input columns were changed, and it
was determined that in order to correctly get the max
amount of reversals made, I needed to use 'RestCount'
which denotes how many trials were left in the task.
When this was 0, that denotes the highest number of
reversals, since everything after is just the program
running through remaining trials once each.